Improved code quality and implemented logic to turn on and off a the radiator

+++ src/main/java/nl/rubenernst/projects/thermostat/processor/domain/ClimateMeasurement.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ClimateMeasurement {
     private final String room;
     private final ClimateType type;
-    private final float value;
+    private final int value;
 }
